requirements.txt

python-socketio
opencv-python
numpy
requests
websocket-client